requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.9.0
PyYAML==6.0.1
eventlet==0.36.1
google-cloud-storage==2.10.0
google-auth==2.23.0
python-dotenv==1.0.0